+++ tests/RepoM.Plugin.Misc.Tests/Configuration/VerifiedDocs/DocsModuleSettingsTests.VerifyChanges.verified.txt
