+++ tests/RepoM.Plugin.Misc.Tests/DefaultAppSettingsTests/DocsAppSettingsTests.AppSettingsDocumentationGeneration_AppSettings.verified.md
