+++ tests/RepoM.Plugin.Misc.Tests/DefaultAppSettingsTests/DocsAppSettingsTests.AppSettingsJsonFileGeneration.verified.md
