+++ tests/RepoM.Plugin.Misc.Tests/TestFramework/NuDoc/ClassWriter.cs
@@ -4,13 +4,7 @@ using System.IO;
 
 internal class ClassWriter
 {
-    public ClassWriter()
-    {
-        Head = new StringWriter();
-        Properties = new StringWriter();
-    }
+    public readonly StringWriter Head = new();
 
-    public readonly StringWriter Head;
-
-    public readonly StringWriter Properties;
+    public readonly StringWriter Properties = new();
 }
